+++ src/Test/MuCheck.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE RecordWildCards #-}
 -- | MuCheck base module
-module Test.MuCheck (mucheck) where
+module Test.MuCheck (mucheck, sampler) where
 
 import Test.MuCheck.Mutation
 import Test.MuCheck.Config